Fixed some bugs in case differentiation in stp7.cc. Changed some method name to make it easy to see, which method needs the time array as intervalls and which needs it as points of time.
Completed the list of tests for checking the case differentiation.

Now all profile with a cruise phase should be computed correctly and for any profile the class should give back the right type. (Only in double decceleration case the first part of the profile can't be determined due to the algorithm. The answer is ?-T or ?-W).